Updated some minor details on statistics
git-svn-id: http://svn.code.sf.net/p/theairline/code/trunk@239 a67f925f-2c50-4282-93e8-9e6206ef1626

+++ TheAirline/GraphicsModel/PageModel/PageAirlineModel/PanelAirlineModel/PageAirlineStatistics.xaml.cs
@@ -100,8 +100,8 @@ namespace TheAirline.GraphicsModel.PageModel.PageAirlineModel.PanelAirlineModel
             }
             else
             {
-                int currentYearValue = sa.Key.Statistics.getStatisticsValue(GameObject.GetInstance().GameTime.Year, sa.Value);
-                int lastYearValue = sa.Key.Statistics.getStatisticsValue(GameObject.GetInstance().GameTime.Year - 1, sa.Value);
+                double currentYearValue = sa.Key.Statistics.getStatisticsValue(GameObject.GetInstance().GameTime.Year, sa.Value);
+                double lastYearValue = sa.Key.Statistics.getStatisticsValue(GameObject.GetInstance().GameTime.Year - 1, sa.Value);
 
                 if (year == 0)
                     return currentYearValue;
